Port concat ops module into rook
