+++ src/rook/utils/ops/__init__.py
@@ -2,6 +2,7 @@
 
 from .base import Operation
 from .average import Average, average_over_dims, average_shape, average_time
+from .concat import Concat, concat
 from .regrid import Regrid, regrid
 from .subset import Subset, subset
 from .normalise import ResultSet, normalise
@@ -12,6 +13,8 @@ __all__ = [
 	"average_over_dims",
 	"average_shape",
 	"average_time",
+	"Concat",
+	"concat",
 	"Regrid",
 	"regrid",
 	"Subset",
